Isolate ingest endpoint (#187)
* Create folder and file to house ingest api route

* Refactor ingest route to be in its own file in ingest folder

* Refactored SESSIONS, chroma_settings, hnsw_metadata to be in config/shared to avoid circular imports

* Refactor is stable

* Delete comments and add import for textwrap

---------
