+++ fastapi-backend/dependencies/llm.py
@@ -1,4 +1,4 @@
-# dependencies/llm.py
+# fastapi-backend/dependencies/llm.py
 import ollama
 # will configure llm to users choices, default is llama3.1
 def get_llm():
