+++ fastapi-backend/dependencies/vectorstore.py
@@ -1,7 +1,9 @@
-# dependencies/vectorstore.py
+# fastapi-backend/dependencies/vectorstore.py
 from fastapi import Body, Request, Depends, HTTPException
 from rag_calls.models import SingleRagRequest
 # from main import VECTOR_STORES
+from config.shared import SESSIONS
+
 
 
 def get_vectorstore(payload: SingleRagRequest = Body(...)):
